+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import ToDoPage from './ToDoPage';
+import ToDoPage from './components/ToDoPage';
 
 import './App.css';
